requirements.txt
flask
google-genai
gunicorn
requests
